standardize v3 adapter structure
