+++ packages/sources/finage-test/src/endpoint/commodities.ts
@@ -1,0 +1,19 @@
+import { PriceEndpoint } from '@chainlink/external-adapter-framework/adapter'
+import { SingleNumberResultResponse } from '@chainlink/external-adapter-framework/util'
+import { config } from '../config'
+import overrides from '../config/overrides.json'
+import { priceInputParameters } from './utils'
+import { transport } from '../transport/commodities'
+
+export type BaseEndpointTypes = {
+  Parameters: typeof priceInputParameters.definition
+  Settings: typeof config.settings
+  Response: SingleNumberResultResponse
+}
+
+export const endpoint = new PriceEndpoint({
+  name: 'commodities',
+  transport,
+  inputParameters: priceInputParameters,
+  overrides: overrides.finage,
+})
