+++ packages/sources/finage-test/src/endpoint/index.ts
@@ -1,6 +1,6 @@
 export { endpoint as crypto } from './crypto-router'
 export { endpoint as stock } from './stock-router'
-export { endpoint as eod } from './http/eod'
-export { endpoint as commodities } from './http/commodities'
+export { endpoint as eod } from './eod'
+export { endpoint as commodities } from './commodities'
 export { endpoint as forex } from './forex-router'
 export { endpoint as ukEtf } from './uk-etf-router'
