+++ packages/sources/finage-test/src/endpoint/stock-router.ts
@@ -1,11 +1,11 @@
 import { AdapterEndpoint } from '@chainlink/external-adapter-framework/adapter'
 import { TransportRoutes } from '@chainlink/external-adapter-framework/transports'
 import overrides from '../config/overrides.json'
-import { httpTransport } from './http/stock'
-import { wsTransport } from './ws/stock-ws'
-import { EquitiesEndpointTypes, equitiesInputParameters } from './types'
+import { httpTransport } from '../transport/stock-http'
+import { wsTransport } from '../transport/stock-ws'
+import { EquitiesEndpointTypes, equitiesInputParameters } from './utils'
 
-export const endpoint = new AdapterEndpoint<EquitiesEndpointTypes>({
+export const endpoint = new AdapterEndpoint({
   name: 'stock',
   transportRoutes: new TransportRoutes<EquitiesEndpointTypes>()
     .register('ws', wsTransport)
