+++ packages/sources/finage-test/src/transport/crypto-ws.ts
@@ -1,5 +1,5 @@
 import { WebsocketReverseMappingTransport } from '@chainlink/external-adapter-framework/transports/websocket'
-import { PriceEndpointTypes } from '../types'
+import { BaseEndpointTypes } from '../endpoint/utils'
 
 interface Message {
   s: string
@@ -8,14 +8,14 @@ interface Message {
   t: number
 }
 
-type EndpointTypes = PriceEndpointTypes & {
+type WsTransportTypes = BaseEndpointTypes & {
   Provider: {
     WsMessage: Message
   }
 }
 
-export const wsTransport: WebsocketReverseMappingTransport<EndpointTypes, string> =
-  new WebsocketReverseMappingTransport<EndpointTypes, string>({
+export const wsTransport: WebsocketReverseMappingTransport<WsTransportTypes, string> =
+  new WebsocketReverseMappingTransport<WsTransportTypes, string>({
     url: (context) => {
       return `${context.adapterSettings.CRYPTO_WS_API_ENDPOINT}/?token=${context.adapterSettings.WS_SOCKET_KEY}`
     },
