+++ packages/sources/finage-test/src/transport/uk-etf-http.ts
@@ -1,14 +1,14 @@
 import { HttpTransport } from '@chainlink/external-adapter-framework/transports'
-import { PriceEndpointTypes } from '../types'
+import { EquitiesEndpointTypes } from '../endpoint/utils'
 
 interface ResponseSchema {
   symbol: string
-  bid: number
-  ask: number
+  price: number
   timestamp: number
+  error?: string
 }
 
-type HttpTransportTypes = PriceEndpointTypes & {
+type HttpTransportTypes = EquitiesEndpointTypes & {
   Provider: {
     RequestBody: never
     ResponseBody: ResponseSchema
@@ -18,27 +18,39 @@ type HttpTransportTypes = PriceEndpointTypes & {
 export const httpTransport = new HttpTransport<HttpTransportTypes>({
   prepareRequests: (params, config) => {
     return params.map((param) => {
-      const symbol = `${param.base}${param.quote}`.toUpperCase()
+      const symbol = param.base?.toUpperCase()
       return {
         params: [param],
         request: {
           baseURL: config.API_ENDPOINT,
-          url: `/last/forex/${symbol}`,
+          url: `/last/etf/${symbol}`,
           params: { apikey: config.API_KEY },
         },
       }
     })
   },
   parseResponse: (params, res) => {
+    if (res.data.error) {
+      return params.map((param) => {
+        return {
+          params: param,
+          response: {
+            errorMessage:
+              "Could not retrieve valid data from Data Provider's /last/etf API. This is likely an issue with the Data Provider or the input params/overrides",
+            statusCode: 400,
+          },
+        }
+      })
+    }
+
     return params.map((param) => {
-      const result = (res.data.ask + res.data.bid) / 2
       return {
         params: param,
         response: {
           data: {
-            result,
+            result: res.data.price,
           },
-          result,
+          result: res.data.price,
           timestamps: {
             providerIndicatedTimeUnixMs: res.data.timestamp,
           },
