+++ packages/sources/finnhub/src/endpoint/quote.ts
@@ -1,11 +1,9 @@
 import { PriceEndpoint } from '@chainlink/external-adapter-framework/adapter'
-import { HttpTransport } from '@chainlink/external-adapter-framework/transports'
-import { SingleNumberResultResponse, makeLogger } from '@chainlink/external-adapter-framework/util'
+import { SingleNumberResultResponse } from '@chainlink/external-adapter-framework/util'
 import { InputParameters } from '@chainlink/external-adapter-framework/validation'
 import { config } from '../config'
 import overrides from '../config/overrides.json'
-
-const logger = makeLogger('Finnhub quote endpoint')
+import { transport } from '../transport/quote'
 
 export const inputParameters = new InputParameters({
   base: {
@@ -21,84 +19,16 @@ export const inputParameters = new InputParameters({
   },
 })
 
-export interface ProviderResponseBody {
-  c: number
-  d: number
-  dp: number
-  h: number
-  l: number
-  o: number
-  pc: number
-  t: number
-}
-
-export interface RequestParams {
-  base: string
-}
-
-export type EndpointTypes = {
+export type BaseEndpointTypes = {
   Parameters: typeof inputParameters.definition
   Settings: typeof config.settings
   Response: SingleNumberResultResponse
-  Provider: {
-    RequestBody: never
-    ResponseBody: ProviderResponseBody
-  }
 }
 
-export const httpTransport = new HttpTransport<EndpointTypes>({
-  prepareRequests: (params, settings: typeof config.settings) => {
-    return params.map((param) => {
-      const symbol = param.base.toUpperCase()
-      const requestConfig = {
-        baseURL: `${settings.API_ENDPOINT}/quote`,
-        method: 'GET',
-        params: {
-          symbol,
-          token: settings.API_KEY,
-        },
-      }
-      return {
-        params: [param],
-        request: requestConfig,
-      }
-    })
-  },
-  parseResponse: (params, res) => {
-    const data = res.data
-    if (!data.c) {
-      return params.map((param) => {
-        const errorMessage = `No data found for ${param.base}`
-        logger.info(errorMessage)
-        return {
-          params: param,
-          response: {
-            statusCode: 502,
-            errorMessage,
-          },
-        }
-      })
-    }
-
-    return params.map((param) => {
-      const result = data.c
-      return {
-        params: param,
-        response: {
-          data: {
-            result,
-          },
-          result,
-        },
-      }
-    })
-  },
-})
-
-export const endpoint = new PriceEndpoint<EndpointTypes>({
+export const endpoint = new PriceEndpoint({
   name: 'quote',
   aliases: ['common', 'stock', 'forex'],
-  transport: httpTransport,
+  transport,
   inputParameters: inputParameters,
   overrides: overrides.finnhub,
 })
