+++ packages/sources/finnhub/src/transport/quote.ts
@@ -1,0 +1,72 @@
+import { config } from '../config'
+import { BaseEndpointTypes } from '../endpoint/quote'
+import { HttpTransport } from '@chainlink/external-adapter-framework/transports'
+import { makeLogger } from '@chainlink/external-adapter-framework/util'
+
+const logger = makeLogger('Finnhub quote endpoint')
+interface ProviderResponseBody {
+  c: number
+  d: number
+  dp: number
+  h: number
+  l: number
+  o: number
+  pc: number
+  t: number
+}
+
+export type HttpTransportTypes = BaseEndpointTypes & {
+  Provider: {
+    RequestBody: never
+    ResponseBody: ProviderResponseBody
+  }
+}
+
+export const transport = new HttpTransport<HttpTransportTypes>({
+  prepareRequests: (params, settings: typeof config.settings) => {
+    return params.map((param) => {
+      const symbol = param.base.toUpperCase()
+      const requestConfig = {
+        baseURL: `${settings.API_ENDPOINT}/quote`,
+        method: 'GET',
+        params: {
+          symbol,
+          token: settings.API_KEY,
+        },
+      }
+      return {
+        params: [param],
+        request: requestConfig,
+      }
+    })
+  },
+  parseResponse: (params, res) => {
+    const data = res.data
+    if (!data.c) {
+      return params.map((param) => {
+        const errorMessage = `No data found for ${param.base}`
+        logger.info(errorMessage)
+        return {
+          params: param,
+          response: {
+            statusCode: 502,
+            errorMessage,
+          },
+        }
+      })
+    }
+
+    return params.map((param) => {
+      const result = data.c
+      return {
+        params: param,
+        response: {
+          data: {
+            result,
+          },
+          result,
+        },
+      }
+    })
+  },
+})
